+++ tecte/Baekjoon/BOJ_0000_template.java
@@ -1,0 +1,21 @@
+/*
+ * 문제 : 백준 0000번 - 0000 (https://www.acmicpc.net/problem/0000)
+ * 문제 풀이
+ *
+ */
+
+package Baekjoon;
+
+import java.io.*;
+
+public class BOJ_0000_template {
+    // public class Main {
+    public static void main(String[] args) throws IOException {
+        BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
+        BufferedWriter bw = new BufferedWriter(new OutputStreamWriter(System.out));
+
+
+        br.close();
+        bw.close();
+    }
+}
